Offsets to int64
These are more properly int64s as they can be very, very large. (This is
what the Kafka client we're using uses, so this makes it slightly
easier/more standard.)

+++ marshal/topic.go
@@ -23,7 +23,7 @@ type topicState struct {
 // PartitionClaim contains claim information about a given partition.
 type PartitionClaim struct {
 	LastHeartbeat int64
-	LastOffset    int
+	LastOffset    int64
 	ClientID      string
 	GroupID       string
